Moved components folder to root of the project
