+++ app/(site)/components/AuthForm.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import Button from "@/app/components/Button";
-import Input from "@/app/components/inputs/Input";
+import Button from "@/components/Button";
+import Input from "@/components/inputs/Input";
 import React, { useCallback, useEffect, useState } from "react";
 import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
 import AuthSocialButton from "./AuthSocialButtont";
