+++ app/conversations/layout.tsx
@@ -1,6 +1,6 @@
 import getConversations from "../../actions/getConversations";
 import getUsers from "../../actions/getUsers";
-import Sidebar from "../components/sidebar/Sidebar";
+import Sidebar from "@/components/sidebar/Sidebar";
 import ConversationList from "./components/ConversationList";
 
 /**
