+++ app/conversations/page.tsx
@@ -2,7 +2,7 @@
 
 import clsx from "clsx";
 import useConversation from "../hooks/useConversation";
-import EmptyState from "../components/EmptyState";
+import EmptyState from "@/components/EmptyState";
 
 /**
  * Renders the conversations page component.
