+++ app/layout.tsx
@@ -1,4 +1,4 @@
-import ActiveStatus from "./components/ActiveStatus";
+import ActiveStatus from "@/components/ActiveStatus";
 import AuthContext from "./context/AuthContext";
 import ToasterContext from "./context/ToasterContext";
 import "./globals.css";
