+++ app/users/layout.tsx
@@ -1,5 +1,5 @@
 import getUsers from "../../actions/getUsers";
-import Sidebar from "../components/sidebar/Sidebar";
+import Sidebar from "@/components/sidebar/Sidebar";
 import UserList from "./components/UserList";
 
 /**
