+++ components/Avatar.tsx
@@ -1,8 +1,8 @@
 "use client";
 
+import useActiveList from "@/app/hooks/useActiveList";
 import { User } from "@prisma/client";
 
-import useActiveList from "../hooks/useActiveList";
 import Image from "next/image";
 
 interface AvatarProps {
