+++ components/modals/ConfirmModal.tsx
@@ -5,8 +5,8 @@ import { Dialog } from "@headlessui/react";
 import { FiAlertTriangle } from "react-icons/fi";
 import axios from "axios";
 import { useRouter } from "next/navigation";
-import Modal from "@/app/components/modals/Modal";
-import Button from "@/app/components/Button";
+import Modal from "@/components/modals/Modal";
+import Button from "@/components/Button";
 import useConversation from "@/app/hooks/useConversation";
 import { toast } from "react-hot-toast";
