+++ components/modals/ImageModal.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Modal from "@/app/components/modals/Modal";
+import Modal from "@/components/modals/Modal";
 import Image from "next/image";
 
 interface ImageModalProps {
