+++ components/sidebar/DesktopSidebar.tsx
@@ -1,9 +1,7 @@
 "use client";
 
 import useRoutes from "@/app/hooks/useRoutes";
-// import SettingsModal from "./SettingsModal";
 import { useState } from "react";
-// import Avatar from "../Avatar";
 import { User } from "@prisma/client";
 import DesktopItem from "./DesktopItem";
 import Avatar from "../Avatar";
