+++ components/sidebar/MobileFooter.tsx
